Module ReactDOM

module Style : sig ... end

val querySelector : string -> Dom.element option

module Client : sig ... end

val render : React.element -> Dom.element -> unit

deprecated This function will be removed in the next release. Use ReactDOM.Client.createRoot and ReactDOM.Client.render instead.

val hydrate : React.element -> Dom.element -> unit

deprecated This function will be removed in the next release. Use ReactDOM.Client.hydrateRoot instead.

val createPortal : React.element -> Dom.element -> React.element

val unmountComponentAtNode : Dom.element -> unit

val flushSync : (unit -> unit) -> unit

val domElementToObj : Dom.element -> < .. > Js.t

type style = Style.t

type domRef

module Ref : sig ... end

type domProps = {

key : string option;

ref : domRef option;

children : React.element option;

ariaActivedescendant : string option;

ariaAtomic : bool option;

ariaAutocomplete : string option;

ariaBusy : bool option;

ariaChecked : string option;

ariaColcount : int option;

ariaColindex : int option;

ariaColspan : int option;

ariaControls : string option;

ariaCurrent : string option;

ariaDescribedby : string option;

ariaDetails : string option;

ariaDisabled : bool option;

ariaErrormessage : string option;

ariaExpanded : bool option;

ariaFlowto : string option;

ariaGrabbed : bool option;

ariaHaspopup : string option;

ariaHidden : bool option;

ariaInvalid : string option;

ariaKeyshortcuts : string option;

ariaLabel : string option;

ariaLabelledby : string option;

ariaLevel : int option;

ariaLive : string option;

ariaModal : bool option;

ariaMultiline : bool option;

ariaMultiselectable : bool option;

ariaOrientation : string option;

ariaOwns : string option;

ariaPlaceholder : string option;

ariaPosinset : int option;

ariaPressed : string option;

ariaReadonly : bool option;

ariaRelevant : string option;

ariaRequired : bool option;

ariaRoledescription : string option;

ariaRowcount : int option;

ariaRowindex : int option;

ariaRowindextext : string option;

ariaRowspan : int option;

ariaSelected : bool option;

ariaSetsize : int option;

ariaSort : string option;

ariaValuemax : float option;

ariaValuemin : float option;

ariaValuenow : float option;

ariaValuetext : string option;

defaultChecked : bool option;

defaultValue : string option;

accessKey : string option;

className : string option;

contentEditable : bool option;

contextMenu : string option;

dir : string option;

draggable : bool option;

hidden : bool option;

id : string option;

lang : string option;

role : string option;

style : style option;

spellCheck : bool option;

tabIndex : int option;

title : string option;

itemID : string option;

itemProp : string option;

itemRef : string option;

itemScope : bool option;

itemType : string option;

as_ : string option;

accept : string option;

acceptCharset : string option;

action : string option;

allowFullScreen : bool option;

alt : string option;

async : bool option;

autoComplete : string option;

autoCapitalize : string option;

autoFocus : bool option;

autoPlay : bool option;

challenge : string option;

charSet : string option;

checked : bool option;

cite : string option;

crossOrigin : string option;

cols : int option;

colSpan : int option;

content : string option;

controls : bool option;

coords : string option;

data : string option;

dateTime : string option;

default : bool option;

defer : bool option;

disabled : bool option;

download : string option;

encType : string option;

form : string option;

formAction : string option;

formTarget : string option;

formMethod : string option;

headers : string option;

height : string option;

high : int option;

href : string option;

hrefLang : string option;

htmlFor : string option;

httpEquiv : string option;

icon : string option;

inputMode : string option;

integrity : string option;

keyType : string option;

kind : string option;

label : string option;

list : string option;

loop : bool option;

low : int option;

manifest : string option;

max : string option;

maxLength : int option;

media : string option;

mediaGroup : string option;

method : string option;

min : string option;

minLength : int option;

multiple : bool option;

muted : bool option;

name : string option;

nonce : string option;

noValidate : bool option;

open_ : bool option;

optimum : int option;

pattern : string option;

placeholder : string option;

playsInline : bool option;

poster : string option;

preload : string option;

radioGroup : string option;

readOnly : bool option;

referrerPolicy : string option;

rel : string option;

required : bool option;

reversed : bool option;

rows : int option;

rowSpan : int option;

sandbox : string option;

scope : string option;

scoped : bool option;

scrolling : string option;

selected : bool option;

shape : string option;

size : int option;

sizes : string option;

span : int option;

src : string option;

srcDoc : string option;

srcLang : string option;

srcSet : string option;

start : int option;

step : float option;

summary : string option;

target : string option;

type_ : string option;

useMap : string option;

value : string option;

width : string option;

wrap : string option;

onCopy : (React.Event.Clipboard.t -> unit) option;

onCut : (React.Event.Clipboard.t -> unit) option;

onPaste : (React.Event.Clipboard.t -> unit) option;

onCompositionEnd : (React.Event.Composition.t -> unit) option;

onCompositionStart : (React.Event.Composition.t -> unit) option;

onCompositionUpdate : (React.Event.Composition.t -> unit) option;

onKeyDown : (React.Event.Keyboard.t -> unit) option;

onKeyPress : (React.Event.Keyboard.t -> unit) option;

onKeyUp : (React.Event.Keyboard.t -> unit) option;

onFocus : (React.Event.Focus.t -> unit) option;

onBlur : (React.Event.Focus.t -> unit) option;

onChange : (React.Event.Form.t -> unit) option;

onInput : (React.Event.Form.t -> unit) option;

onSubmit : (React.Event.Form.t -> unit) option;

onInvalid : (React.Event.Form.t -> unit) option;

onClick : (React.Event.Mouse.t -> unit) option;

onContextMenu : (React.Event.Mouse.t -> unit) option;

onDoubleClick : (React.Event.Mouse.t -> unit) option;

onDrag : (React.Event.Drag.t -> unit) option;

onDragEnd : (React.Event.Drag.t -> unit) option;

onDragEnter : (React.Event.Drag.t -> unit) option;

onDragExit : (React.Event.Drag.t -> unit) option;

onDragLeave : (React.Event.Drag.t -> unit) option;

onDragOver : (React.Event.Drag.t -> unit) option;

onDragStart : (React.Event.Drag.t -> unit) option;

onDrop : (React.Event.Drag.t -> unit) option;

onMouseDown : (React.Event.Mouse.t -> unit) option;

onMouseEnter : (React.Event.Mouse.t -> unit) option;

onMouseLeave : (React.Event.Mouse.t -> unit) option;

onMouseMove : (React.Event.Mouse.t -> unit) option;

onMouseOut : (React.Event.Mouse.t -> unit) option;

onMouseOver : (React.Event.Mouse.t -> unit) option;

onMouseUp : (React.Event.Mouse.t -> unit) option;

onSelect : (React.Event.Selection.t -> unit) option;

onTouchCancel : (React.Event.Touch.t -> unit) option;

onTouchEnd : (React.Event.Touch.t -> unit) option;

onTouchMove : (React.Event.Touch.t -> unit) option;

onTouchStart : (React.Event.Touch.t -> unit) option;

onPointerOver : (React.Event.Pointer.t -> unit) option;

onPointerEnter : (React.Event.Pointer.t -> unit) option;

onPointerDown : (React.Event.Pointer.t -> unit) option;

onPointerMove : (React.Event.Pointer.t -> unit) option;

onPointerUp : (React.Event.Pointer.t -> unit) option;

onPointerCancel : (React.Event.Pointer.t -> unit) option;

onPointerOut : (React.Event.Pointer.t -> unit) option;

onPointerLeave : (React.Event.Pointer.t -> unit) option;

onGotPointerCapture : (React.Event.Pointer.t -> unit) option;

onLostPointerCapture : (React.Event.Pointer.t -> unit) option;

onScroll : (React.Event.UI.t -> unit) option;

onWheel : (React.Event.Wheel.t -> unit) option;

onAbort : (React.Event.Media.t -> unit) option;

onCanPlay : (React.Event.Media.t -> unit) option;

onCanPlayThrough : (React.Event.Media.t -> unit) option;

onDurationChange : (React.Event.Media.t -> unit) option;

onEmptied : (React.Event.Media.t -> unit) option;

onEncrypetd : (React.Event.Media.t -> unit) option;

onEnded : (React.Event.Media.t -> unit) option;

onError : (React.Event.Media.t -> unit) option;

onLoadedData : (React.Event.Media.t -> unit) option;

onLoadedMetadata : (React.Event.Media.t -> unit) option;

onLoadStart : (React.Event.Media.t -> unit) option;

onPause : (React.Event.Media.t -> unit) option;

onPlay : (React.Event.Media.t -> unit) option;

onPlaying : (React.Event.Media.t -> unit) option;

onProgress : (React.Event.Media.t -> unit) option;

onRateChange : (React.Event.Media.t -> unit) option;

onSeeked : (React.Event.Media.t -> unit) option;

onSeeking : (React.Event.Media.t -> unit) option;

onStalled : (React.Event.Media.t -> unit) option;

onSuspend : (React.Event.Media.t -> unit) option;

onTimeUpdate : (React.Event.Media.t -> unit) option;

onVolumeChange : (React.Event.Media.t -> unit) option;

onWaiting : (React.Event.Media.t -> unit) option;

onLoad : (React.Event.Image.t -> unit) option;

onAnimationStart : (React.Event.Animation.t -> unit) option;

onAnimationEnd : (React.Event.Animation.t -> unit) option;

onAnimationIteration : (React.Event.Animation.t -> unit) option;

onTransitionEnd : (React.Event.Transition.t -> unit) option;

accentHeight : string option;

accumulate : string option;

additive : string option;

alignmentBaseline : string option;

allowReorder : string option;

alphabetic : string option;

amplitude : string option;

arabicForm : string option;

ascent : string option;

attributeName : string option;

attributeType : string option;

autoReverse : string option;

azimuth : string option;

baseFrequency : string option;

baseProfile : string option;

baselineShift : string option;

bbox : string option;

begin_ : string option;

bias : string option;

by : string option;

calcMode : string option;

capHeight : string option;

clip : string option;

clipPath : string option;

clipPathUnits : string option;

clipRule : string option;

colorInterpolation : string option;

colorInterpolationFilters : string option;

colorProfile : string option;

colorRendering : string option;

contentScriptType : string option;

contentStyleType : string option;

cursor : string option;

cx : string option;

cy : string option;

d : string option;

decelerate : string option;

descent : string option;

diffuseConstant : string option;

direction : string option;

display : string option;

divisor : string option;

dominantBaseline : string option;

dur : string option;

dx : string option;

dy : string option;

edgeMode : string option;

elevation : string option;

enableBackground : string option;

end_ : string option;

exponent : string option;

externalResourcesRequired : string option;

fill : string option;

fillOpacity : string option;

fillRule : string option;

filter : string option;

filterRes : string option;

filterUnits : string option;

floodColor : string option;

floodOpacity : string option;

focusable : string option;

fontFamily : string option;

fontSize : string option;

fontSizeAdjust : string option;

fontStretch : string option;

fontStyle : string option;

fontVariant : string option;

fontWeight : string option;

fomat : string option;

from : string option;

fx : string option;

fy : string option;

g1 : string option;

g2 : string option;

glyphName : string option;

glyphOrientationHorizontal : string option;

glyphOrientationVertical : string option;

glyphRef : string option;

gradientTransform : string option;

gradientUnits : string option;

hanging : string option;

horizAdvX : string option;

horizOriginX : string option;

ideographic : string option;

imageRendering : string option;

in_ : string option;

in2 : string option;

intercept : string option;

k : string option;

k1 : string option;

k2 : string option;

k3 : string option;

k4 : string option;

kernelMatrix : string option;

kernelUnitLength : string option;

kerning : string option;

keyPoints : string option;

keySplines : string option;

keyTimes : string option;

lengthAdjust : string option;

letterSpacing : string option;

lightingColor : string option;

limitingConeAngle : string option;

local : string option;

markerEnd : string option;

markerHeight : string option;

markerMid : string option;

markerStart : string option;

markerUnits : string option;

markerWidth : string option;

mask : string option;

maskContentUnits : string option;

maskUnits : string option;

mathematical : string option;

mode : string option;

numOctaves : string option;

offset : string option;

opacity : string option;

operator : string option;

order : string option;

orient : string option;

orientation : string option;

origin : string option;

overflow : string option;

overflowX : string option;

overflowY : string option;

overlinePosition : string option;

overlineThickness : string option;

paintOrder : string option;

panose1 : string option;

pathLength : string option;

patternContentUnits : string option;

patternTransform : string option;

patternUnits : string option;

pointerEvents : string option;

points : string option;

pointsAtX : string option;

pointsAtY : string option;

pointsAtZ : string option;

preserveAlpha : string option;

preserveAspectRatio : string option;

primitiveUnits : string option;

r : string option;

radius : string option;

refX : string option;

refY : string option;

renderingIntent : string option;

repeatCount : string option;

repeatDur : string option;

requiredExtensions : string option;

requiredFeatures : string option;

restart : string option;

result : string option;

rotate : string option;

rx : string option;

ry : string option;

scale : string option;

seed : string option;

shapeRendering : string option;

slope : string option;

spacing : string option;

specularConstant : string option;

specularExponent : string option;

speed : string option;

spreadMethod : string option;

startOffset : string option;

stdDeviation : string option;

stemh : string option;

stemv : string option;

stitchTiles : string option;

stopColor : string option;

stopOpacity : string option;

strikethroughPosition : string option;

strikethroughThickness : string option;

string : string option;

stroke : string option;

strokeDasharray : string option;

strokeDashoffset : string option;

strokeLinecap : string option;

strokeLinejoin : string option;

strokeMiterlimit : string option;

strokeOpacity : string option;

strokeWidth : string option;

surfaceScale : string option;

systemLanguage : string option;

tableValues : string option;

targetX : string option;

targetY : string option;

textAnchor : string option;

textDecoration : string option;

textLength : string option;

textRendering : string option;

to_ : string option;

transform : string option;

u1 : string option;

u2 : string option;

underlinePosition : string option;

underlineThickness : string option;

unicode : string option;

unicodeBidi : string option;

unicodeRange : string option;

unitsPerEm : string option;

vAlphabetic : string option;

vHanging : string option;

vIdeographic : string option;

vMathematical : string option;

values : string option;

vectorEffect : string option;

version : string option;

vertAdvX : string option;

vertAdvY : string option;

vertOriginX : string option;

vertOriginY : string option;

viewBox : string option;

viewTarget : string option;

visibility : string option;

widths : string option;

wordSpacing : string option;

writingMode : string option;

x : string option;

x1 : string option;

x2 : string option;

xChannelSelector : string option;

xHeight : string option;

xlinkActuate : string option;

xlinkArcrole : string option;

xlinkHref : string option;

xlinkRole : string option;

xlinkShow : string option;

xlinkTitle : string option;

xlinkType : string option;

xmlns : string option;

xmlnsXlink : string option;

xmlBase : string option;

xmlLang : string option;

xmlSpace : string option;

y : string option;

y1 : string option;

y2 : string option;

yChannelSelector : string option;

z : string option;

zoomAndPan : string option;

about : string option;

datatype : string option;

inlist : string option;

prefix : string option;

property : string option;

resource : string option;

typeof : string option;

vocab : string option;

dangerouslySetInnerHTML : < __html : string > Js.t option;

suppressContentEditableWarning : bool option;

suppressHydrationWarning : bool option;

}

val domProps : ?key:string -> ?ref:domRef -> ?children:React.element -> ?ariaActivedescendant:string -> ?ariaAtomic:bool -> ?ariaAutocomplete:string -> ?ariaBusy:bool -> ?ariaChecked:string -> ?ariaColcount:int -> ?ariaColindex:int -> ?ariaColspan:int -> ?ariaControls:string -> ?ariaCurrent:string -> ?ariaDescribedby:string -> ?ariaDetails:string -> ?ariaDisabled:bool -> ?ariaErrormessage:string -> ?ariaExpanded:bool -> ?ariaFlowto:string -> ?ariaGrabbed:bool -> ?ariaHaspopup:string -> ?ariaHidden:bool -> ?ariaInvalid:string -> ?ariaKeyshortcuts:string -> ?ariaLabel:string -> ?ariaLabelledby:string -> ?ariaLevel:int -> ?ariaLive:string -> ?ariaModal:bool -> ?ariaMultiline:bool -> ?ariaMultiselectable:bool -> ?ariaOrientation:string -> ?ariaOwns:string -> ?ariaPlaceholder:string -> ?ariaPosinset:int -> ?ariaPressed:string -> ?ariaReadonly:bool -> ?ariaRelevant:string -> ?ariaRequired:bool -> ?ariaRoledescription:string -> ?ariaRowcount:int -> ?ariaRowindex:int -> ?ariaRowindextext:string -> ?ariaRowspan:int -> ?ariaSelected:bool -> ?ariaSetsize:int -> ?ariaSort:string -> ?ariaValuemax:float -> ?ariaValuemin:float -> ?ariaValuenow:float -> ?ariaValuetext:string -> ?defaultChecked:bool -> ?defaultValue:string -> ?accessKey:string -> ?className:string -> ?contentEditable:bool -> ?contextMenu:string -> ?dir:string -> ?draggable:bool -> ?hidden:bool -> ?id:string -> ?lang:string -> ?role:string -> ?style:style -> ?spellCheck:bool -> ?tabIndex:int -> ?title:string -> ?itemID:string -> ?itemProp:string -> ?itemRef:string -> ?itemScope:bool -> ?itemType:string -> ?as_:string -> ?accept:string -> ?acceptCharset:string -> ?action:string -> ?allowFullScreen:bool -> ?alt:string -> ?async:bool -> ?autoComplete:string -> ?autoCapitalize:string -> ?autoFocus:bool -> ?autoPlay:bool -> ?challenge:string -> ?charSet:string -> ?checked:bool -> ?cite:string -> ?crossOrigin:string -> ?cols:int -> ?colSpan:int -> ?content:string -> ?controls:bool -> ?coords:string -> ?data:string -> ?dateTime:string -> ?default:bool -> ?defer:bool -> ?disabled:bool -> ?download:string -> ?encType:string -> ?form:string -> ?formAction:string -> ?formTarget:string -> ?formMethod:string -> ?headers:string -> ?height:string -> ?high:int -> ?href:string -> ?hrefLang:string -> ?htmlFor:string -> ?httpEquiv:string -> ?icon:string -> ?inputMode:string -> ?integrity:string -> ?keyType:string -> ?kind:string -> ?label:string -> ?list:string -> ?loop:bool -> ?low:int -> ?manifest:string -> ?max:string -> ?maxLength:int -> ?media:string -> ?mediaGroup:string -> ?method:string -> ?min:string -> ?minLength:int -> ?multiple:bool -> ?muted:bool -> ?name:string -> ?nonce:string -> ?noValidate:bool -> ?open_:bool -> ?optimum:int -> ?pattern:string -> ?placeholder:string -> ?playsInline:bool -> ?poster:string -> ?preload:string -> ?radioGroup:string -> ?readOnly:bool -> ?referrerPolicy:string -> ?rel:string -> ?required:bool -> ?reversed:bool -> ?rows:int -> ?rowSpan:int -> ?sandbox:string -> ?scope:string -> ?scoped:bool -> ?scrolling:string -> ?selected:bool -> ?shape:string -> ?size:int -> ?sizes:string -> ?span:int -> ?src:string -> ?srcDoc:string -> ?srcLang:string -> ?srcSet:string -> ?start:int -> ?step:float -> ?summary:string -> ?target:string -> ?type_:string -> ?useMap:string -> ?value:string -> ?width:string -> ?wrap:string -> ?onCopy:(React.Event.Clipboard.t -> unit) -> ?onCut:(React.Event.Clipboard.t -> unit) -> ?onPaste:(React.Event.Clipboard.t -> unit) -> ?onCompositionEnd:(React.Event.Composition.t -> unit) -> ?onCompositionStart:(React.Event.Composition.t -> unit) -> ?onCompositionUpdate:(React.Event.Composition.t -> unit) -> ?onKeyDown:(React.Event.Keyboard.t -> unit) -> ?onKeyPress:(React.Event.Keyboard.t -> unit) -> ?onKeyUp:(React.Event.Keyboard.t -> unit) -> ?onFocus:(React.Event.Focus.t -> unit) -> ?onBlur:(React.Event.Focus.t -> unit) -> ?onChange:(React.Event.Form.t -> unit) -> ?onInput:(React.Event.Form.t -> unit) -> ?onSubmit:(React.Event.Form.t -> unit) -> ?onInvalid:(React.Event.Form.t -> unit) -> ?onClick:(React.Event.Mouse.t -> unit) -> ?onContextMenu:(React.Event.Mouse.t -> unit) -> ?onDoubleClick:(React.Event.Mouse.t -> unit) -> ?onDrag:(React.Event.Drag.t -> unit) -> ?onDragEnd:(React.Event.Drag.t -> unit) -> ?onDragEnter:(React.Event.Drag.t -> unit) -> ?onDragExit:(React.Event.Drag.t -> unit) -> ?onDragLeave:(React.Event.Drag.t -> unit) -> ?onDragOver:(React.Event.Drag.t -> unit) -> ?onDragStart:(React.Event.Drag.t -> unit) -> ?onDrop:(React.Event.Drag.t -> unit) -> ?onMouseDown:(React.Event.Mouse.t -> unit) -> ?onMouseEnter:(React.Event.Mouse.t -> unit) -> ?onMouseLeave:(React.Event.Mouse.t -> unit) -> ?onMouseMove:(React.Event.Mouse.t -> unit) -> ?onMouseOut:(React.Event.Mouse.t -> unit) -> ?onMouseOver:(React.Event.Mouse.t -> unit) -> ?onMouseUp:(React.Event.Mouse.t -> unit) -> ?onSelect:(React.Event.Selection.t -> unit) -> ?onTouchCancel:(React.Event.Touch.t -> unit) -> ?onTouchEnd:(React.Event.Touch.t -> unit) -> ?onTouchMove:(React.Event.Touch.t -> unit) -> ?onTouchStart:(React.Event.Touch.t -> unit) -> ?onPointerOver:(React.Event.Pointer.t -> unit) -> ?onPointerEnter:(React.Event.Pointer.t -> unit) -> ?onPointerDown:(React.Event.Pointer.t -> unit) -> ?onPointerMove:(React.Event.Pointer.t -> unit) -> ?onPointerUp:(React.Event.Pointer.t -> unit) -> ?onPointerCancel:(React.Event.Pointer.t -> unit) -> ?onPointerOut:(React.Event.Pointer.t -> unit) -> ?onPointerLeave:(React.Event.Pointer.t -> unit) -> ?onGotPointerCapture:(React.Event.Pointer.t -> unit) -> ?onLostPointerCapture:(React.Event.Pointer.t -> unit) -> ?onScroll:(React.Event.UI.t -> unit) -> ?onWheel:(React.Event.Wheel.t -> unit) -> ?onAbort:(React.Event.Media.t -> unit) -> ?onCanPlay:(React.Event.Media.t -> unit) -> ?onCanPlayThrough:(React.Event.Media.t -> unit) -> ?onDurationChange:(React.Event.Media.t -> unit) -> ?onEmptied:(React.Event.Media.t -> unit) -> ?onEncrypetd:(React.Event.Media.t -> unit) -> ?onEnded:(React.Event.Media.t -> unit) -> ?onError:(React.Event.Media.t -> unit) -> ?onLoadedData:(React.Event.Media.t -> unit) -> ?onLoadedMetadata:(React.Event.Media.t -> unit) -> ?onLoadStart:(React.Event.Media.t -> unit) -> ?onPause:(React.Event.Media.t -> unit) -> ?onPlay:(React.Event.Media.t -> unit) -> ?onPlaying:(React.Event.Media.t -> unit) -> ?onProgress:(React.Event.Media.t -> unit) -> ?onRateChange:(React.Event.Media.t -> unit) -> ?onSeeked:(React.Event.Media.t -> unit) -> ?onSeeking:(React.Event.Media.t -> unit) -> ?onStalled:(React.Event.Media.t -> unit) -> ?onSuspend:(React.Event.Media.t -> unit) -> ?onTimeUpdate:(React.Event.Media.t -> unit) -> ?onVolumeChange:(React.Event.Media.t -> unit) -> ?onWaiting:(React.Event.Media.t -> unit) -> ?onLoad:(React.Event.Image.t -> unit) -> ?onAnimationStart:(React.Event.Animation.t -> unit) -> ?onAnimationEnd:(React.Event.Animation.t -> unit) -> ?onAnimationIteration:(React.Event.Animation.t -> unit) -> ?onTransitionEnd:(React.Event.Transition.t -> unit) -> ?accentHeight:string -> ?accumulate:string -> ?additive:string -> ?alignmentBaseline:string -> ?allowReorder:string -> ?alphabetic:string -> ?amplitude:string -> ?arabicForm:string -> ?ascent:string -> ?attributeName:string -> ?attributeType:string -> ?autoReverse:string -> ?azimuth:string -> ?baseFrequency:string -> ?baseProfile:string -> ?baselineShift:string -> ?bbox:string -> ?begin_:string -> ?bias:string -> ?by:string -> ?calcMode:string -> ?capHeight:string -> ?clip:string -> ?clipPath:string -> ?clipPathUnits:string -> ?clipRule:string -> ?colorInterpolation:string -> ?colorInterpolationFilters:string -> ?colorProfile:string -> ?colorRendering:string -> ?contentScriptType:string -> ?contentStyleType:string -> ?cursor:string -> ?cx:string -> ?cy:string -> ?d:string -> ?decelerate:string -> ?descent:string -> ?diffuseConstant:string -> ?direction:string -> ?display:string -> ?divisor:string -> ?dominantBaseline:string -> ?dur:string -> ?dx:string -> ?dy:string -> ?edgeMode:string -> ?elevation:string -> ?enableBackground:string -> ?end_:string -> ?exponent:string -> ?externalResourcesRequired:string -> ?fill:string -> ?fillOpacity:string -> ?fillRule:string -> ?filter:string -> ?filterRes:string -> ?filterUnits:string -> ?floodColor:string -> ?floodOpacity:string -> ?focusable:string -> ?fontFamily:string -> ?fontSize:string -> ?fontSizeAdjust:string -> ?fontStretch:string -> ?fontStyle:string -> ?fontVariant:string -> ?fontWeight:string -> ?fomat:string -> ?from:string -> ?fx:string -> ?fy:string -> ?g1:string -> ?g2:string -> ?glyphName:string -> ?glyphOrientationHorizontal:string -> ?glyphOrientationVertical:string -> ?glyphRef:string -> ?gradientTransform:string -> ?gradientUnits:string -> ?hanging:string -> ?horizAdvX:string -> ?horizOriginX:string -> ?ideographic:string -> ?imageRendering:string -> ?in_:string -> ?in2:string -> ?intercept:string -> ?k:string -> ?k1:string -> ?k2:string -> ?k3:string -> ?k4:string -> ?kernelMatrix:string -> ?kernelUnitLength:string -> ?kerning:string -> ?keyPoints:string -> ?keySplines:string -> ?keyTimes:string -> ?lengthAdjust:string -> ?letterSpacing:string -> ?lightingColor:string -> ?limitingConeAngle:string -> ?local:string -> ?markerEnd:string -> ?markerHeight:string -> ?markerMid:string -> ?markerStart:string -> ?markerUnits:string -> ?markerWidth:string -> ?mask:string -> ?maskContentUnits:string -> ?maskUnits:string -> ?mathematical:string -> ?mode:string -> ?numOctaves:string -> ?offset:string -> ?opacity:string -> ?operator:string -> ?order:string -> ?orient:string -> ?orientation:string -> ?origin:string -> ?overflow:string -> ?overflowX:string -> ?overflowY:string -> ?overlinePosition:string -> ?overlineThickness:string -> ?paintOrder:string -> ?panose1:string -> ?pathLength:string -> ?patternContentUnits:string -> ?patternTransform:string -> ?patternUnits:string -> ?pointerEvents:string -> ?points:string -> ?pointsAtX:string -> ?pointsAtY:string -> ?pointsAtZ:string -> ?preserveAlpha:string -> ?preserveAspectRatio:string -> ?primitiveUnits:string -> ?r:string -> ?radius:string -> ?refX:string -> ?refY:string -> ?renderingIntent:string -> ?repeatCount:string -> ?repeatDur:string -> ?requiredExtensions:string -> ?requiredFeatures:string -> ?restart:string -> ?result:string -> ?rotate:string -> ?rx:string -> ?ry:string -> ?scale:string -> ?seed:string -> ?shapeRendering:string -> ?slope:string -> ?spacing:string -> ?specularConstant:string -> ?specularExponent:string -> ?speed:string -> ?spreadMethod:string -> ?startOffset:string -> ?stdDeviation:string -> ?stemh:string -> ?stemv:string -> ?stitchTiles:string -> ?stopColor:string -> ?stopOpacity:string -> ?strikethroughPosition:string -> ?strikethroughThickness:string -> ?string:string -> ?stroke:string -> ?strokeDasharray:string -> ?strokeDashoffset:string -> ?strokeLinecap:string -> ?strokeLinejoin:string -> ?strokeMiterlimit:string -> ?strokeOpacity:string -> ?strokeWidth:string -> ?surfaceScale:string -> ?systemLanguage:string -> ?tableValues:string -> ?targetX:string -> ?targetY:string -> ?textAnchor:string -> ?textDecoration:string -> ?textLength:string -> ?textRendering:string -> ?to_:string -> ?transform:string -> ?u1:string -> ?u2:string -> ?underlinePosition:string -> ?underlineThickness:string -> ?unicode:string -> ?unicodeBidi:string -> ?unicodeRange:string -> ?unitsPerEm:string -> ?vAlphabetic:string -> ?vHanging:string -> ?vIdeographic:string -> ?vMathematical:string -> ?values:string -> ?vectorEffect:string -> ?version:string -> ?vertAdvX:string -> ?vertAdvY:string -> ?vertOriginX:string -> ?vertOriginY:string -> ?viewBox:string -> ?viewTarget:string -> ?visibility:string -> ?widths:string -> ?wordSpacing:string -> ?writingMode:string -> ?x:string -> ?x1:string -> ?x2:string -> ?xChannelSelector:string -> ?xHeight:string -> ?xlinkActuate:string -> ?xlinkArcrole:string -> ?xlinkHref:string -> ?xlinkRole:string -> ?xlinkShow:string -> ?xlinkTitle:string -> ?xlinkType:string -> ?xmlns:string -> ?xmlnsXlink:string -> ?xmlBase:string -> ?xmlLang:string -> ?xmlSpace:string -> ?y:string -> ?y1:string -> ?y2:string -> ?yChannelSelector:string -> ?z:string -> ?zoomAndPan:string -> ?about:string -> ?datatype:string -> ?inlist:string -> ?prefix:string -> ?property:string -> ?resource:string -> ?typeof:string -> ?vocab:string -> ?dangerouslySetInnerHTML:< __html : string > Js.t -> ?suppressContentEditableWarning:bool -> ?suppressHydrationWarning:bool -> unit -> domProps

val keyGet : domProps -> string option

val refGet : domProps -> domRef option

val childrenGet : domProps -> React.element option

val ariaActivedescendantGet : domProps -> string option

val ariaAtomicGet : domProps -> bool option

val ariaAutocompleteGet : domProps -> string option

val ariaBusyGet : domProps -> bool option

val ariaCheckedGet : domProps -> string option

val ariaColcountGet : domProps -> int option

val ariaColindexGet : domProps -> int option

val ariaColspanGet : domProps -> int option

val ariaControlsGet : domProps -> string option

val ariaCurrentGet : domProps -> string option

val ariaDescribedbyGet : domProps -> string option

val ariaDetailsGet : domProps -> string option

val ariaDisabledGet : domProps -> bool option

val ariaErrormessageGet : domProps -> string option

val ariaExpandedGet : domProps -> bool option

val ariaFlowtoGet : domProps -> string option

val ariaGrabbedGet : domProps -> bool option

val ariaHaspopupGet : domProps -> string option

val ariaHiddenGet : domProps -> bool option

val ariaInvalidGet : domProps -> string option

val ariaKeyshortcutsGet : domProps -> string option

val ariaLabelGet : domProps -> string option

val ariaLabelledbyGet : domProps -> string option

val ariaLevelGet : domProps -> int option

val ariaLiveGet : domProps -> string option

val ariaModalGet : domProps -> bool option

val ariaMultilineGet : domProps -> bool option

val ariaMultiselectableGet : domProps -> bool option

val ariaOrientationGet : domProps -> string option

val ariaOwnsGet : domProps -> string option

val ariaPlaceholderGet : domProps -> string option

val ariaPosinsetGet : domProps -> int option

val ariaPressedGet : domProps -> string option

val ariaReadonlyGet : domProps -> bool option

val ariaRelevantGet : domProps -> string option

val ariaRequiredGet : domProps -> bool option

val ariaRoledescriptionGet : domProps -> string option

val ariaRowcountGet : domProps -> int option

val ariaRowindexGet : domProps -> int option

val ariaRowindextextGet : domProps -> string option

val ariaRowspanGet : domProps -> int option

val ariaSelectedGet : domProps -> bool option

val ariaSetsizeGet : domProps -> int option

val ariaSortGet : domProps -> string option

val ariaValuemaxGet : domProps -> float option

val ariaValueminGet : domProps -> float option

val ariaValuenowGet : domProps -> float option

val ariaValuetextGet : domProps -> string option

val defaultCheckedGet : domProps -> bool option

val defaultValueGet : domProps -> string option

val accessKeyGet : domProps -> string option

val classNameGet : domProps -> string option

val contentEditableGet : domProps -> bool option

val contextMenuGet : domProps -> string option

val dirGet : domProps -> string option

val draggableGet : domProps -> bool option

val hiddenGet : domProps -> bool option

val idGet : domProps -> string option

val langGet : domProps -> string option

val roleGet : domProps -> string option

val styleGet : domProps -> style option

val spellCheckGet : domProps -> bool option

val tabIndexGet : domProps -> int option

val titleGet : domProps -> string option

val itemIDGet : domProps -> string option

val itemPropGet : domProps -> string option

val itemRefGet : domProps -> string option

val itemScopeGet : domProps -> bool option

val itemTypeGet : domProps -> string option

val as_Get : domProps -> string option

val acceptGet : domProps -> string option

val acceptCharsetGet : domProps -> string option

val actionGet : domProps -> string option

val allowFullScreenGet : domProps -> bool option

val altGet : domProps -> string option

val asyncGet : domProps -> bool option

val autoCompleteGet : domProps -> string option

val autoCapitalizeGet : domProps -> string option

val autoFocusGet : domProps -> bool option

val autoPlayGet : domProps -> bool option

val challengeGet : domProps -> string option

val charSetGet : domProps -> string option

val checkedGet : domProps -> bool option

val citeGet : domProps -> string option

val crossOriginGet : domProps -> string option

val colsGet : domProps -> int option

val colSpanGet : domProps -> int option

val contentGet : domProps -> string option

val controlsGet : domProps -> bool option

val coordsGet : domProps -> string option

val dataGet : domProps -> string option

val dateTimeGet : domProps -> string option

val defaultGet : domProps -> bool option

val deferGet : domProps -> bool option

val disabledGet : domProps -> bool option

val downloadGet : domProps -> string option

val encTypeGet : domProps -> string option

val formGet : domProps -> string option

val formActionGet : domProps -> string option

val formTargetGet : domProps -> string option

val formMethodGet : domProps -> string option

val headersGet : domProps -> string option

val heightGet : domProps -> string option

val highGet : domProps -> int option

val hrefGet : domProps -> string option

val hrefLangGet : domProps -> string option

val htmlForGet : domProps -> string option

val httpEquivGet : domProps -> string option

val iconGet : domProps -> string option

val inputModeGet : domProps -> string option

val integrityGet : domProps -> string option

val keyTypeGet : domProps -> string option

val kindGet : domProps -> string option

val labelGet : domProps -> string option

val listGet : domProps -> string option

val loopGet : domProps -> bool option

val lowGet : domProps -> int option

val manifestGet : domProps -> string option

val maxGet : domProps -> string option

val maxLengthGet : domProps -> int option

val mediaGet : domProps -> string option

val mediaGroupGet : domProps -> string option

val methodGet : domProps -> string option

val minGet : domProps -> string option

val minLengthGet : domProps -> int option

val multipleGet : domProps -> bool option

val mutedGet : domProps -> bool option

val nameGet : domProps -> string option

val nonceGet : domProps -> string option

val noValidateGet : domProps -> bool option

val open_Get : domProps -> bool option

val optimumGet : domProps -> int option

val patternGet : domProps -> string option

val placeholderGet : domProps -> string option

val playsInlineGet : domProps -> bool option

val posterGet : domProps -> string option

val preloadGet : domProps -> string option

val radioGroupGet : domProps -> string option

val readOnlyGet : domProps -> bool option

val referrerPolicyGet : domProps -> string option

val relGet : domProps -> string option

val requiredGet : domProps -> bool option

val reversedGet : domProps -> bool option

val rowsGet : domProps -> int option

val rowSpanGet : domProps -> int option

val sandboxGet : domProps -> string option

val scopeGet : domProps -> string option

val scopedGet : domProps -> bool option

val scrollingGet : domProps -> string option

val selectedGet : domProps -> bool option

val shapeGet : domProps -> string option

val sizeGet : domProps -> int option

val sizesGet : domProps -> string option

val spanGet : domProps -> int option

val srcGet : domProps -> string option

val srcDocGet : domProps -> string option

val srcLangGet : domProps -> string option

val srcSetGet : domProps -> string option

val startGet : domProps -> int option

val stepGet : domProps -> float option

val summaryGet : domProps -> string option

val targetGet : domProps -> string option

val type_Get : domProps -> string option

val useMapGet : domProps -> string option

val valueGet : domProps -> string option

val widthGet : domProps -> string option

val wrapGet : domProps -> string option

val onCopyGet : domProps -> (React.Event.Clipboard.t -> unit) option

val onCutGet : domProps -> (React.Event.Clipboard.t -> unit) option

val onPasteGet : domProps -> (React.Event.Clipboard.t -> unit) option

val onCompositionEndGet : domProps -> (React.Event.Composition.t -> unit) option

val onCompositionStartGet : domProps -> (React.Event.Composition.t -> unit) option

val onCompositionUpdateGet : domProps -> (React.Event.Composition.t -> unit) option

val onKeyDownGet : domProps -> (React.Event.Keyboard.t -> unit) option

val onKeyPressGet : domProps -> (React.Event.Keyboard.t -> unit) option

val onKeyUpGet : domProps -> (React.Event.Keyboard.t -> unit) option

val onFocusGet : domProps -> (React.Event.Focus.t -> unit) option

val onBlurGet : domProps -> (React.Event.Focus.t -> unit) option

val onChangeGet : domProps -> (React.Event.Form.t -> unit) option

val onInputGet : domProps -> (React.Event.Form.t -> unit) option

val onSubmitGet : domProps -> (React.Event.Form.t -> unit) option

val onInvalidGet : domProps -> (React.Event.Form.t -> unit) option

val onClickGet : domProps -> (React.Event.Mouse.t -> unit) option

val onContextMenuGet : domProps -> (React.Event.Mouse.t -> unit) option

val onDoubleClickGet : domProps -> (React.Event.Mouse.t -> unit) option

val onDragGet : domProps -> (React.Event.Drag.t -> unit) option

val onDragEndGet : domProps -> (React.Event.Drag.t -> unit) option

val onDragEnterGet : domProps -> (React.Event.Drag.t -> unit) option

val onDragExitGet : domProps -> (React.Event.Drag.t -> unit) option

val onDragLeaveGet : domProps -> (React.Event.Drag.t -> unit) option

val onDragOverGet : domProps -> (React.Event.Drag.t -> unit) option

val onDragStartGet : domProps -> (React.Event.Drag.t -> unit) option

val onDropGet : domProps -> (React.Event.Drag.t -> unit) option

val onMouseDownGet : domProps -> (React.Event.Mouse.t -> unit) option

val onMouseEnterGet : domProps -> (React.Event.Mouse.t -> unit) option

val onMouseLeaveGet : domProps -> (React.Event.Mouse.t -> unit) option

val onMouseMoveGet : domProps -> (React.Event.Mouse.t -> unit) option

val onMouseOutGet : domProps -> (React.Event.Mouse.t -> unit) option

val onMouseOverGet : domProps -> (React.Event.Mouse.t -> unit) option

val onMouseUpGet : domProps -> (React.Event.Mouse.t -> unit) option

val onSelectGet : domProps -> (React.Event.Selection.t -> unit) option

val onTouchCancelGet : domProps -> (React.Event.Touch.t -> unit) option

val onTouchEndGet : domProps -> (React.Event.Touch.t -> unit) option

val onTouchMoveGet : domProps -> (React.Event.Touch.t -> unit) option

val onTouchStartGet : domProps -> (React.Event.Touch.t -> unit) option

val onPointerOverGet : domProps -> (React.Event.Pointer.t -> unit) option

val onPointerEnterGet : domProps -> (React.Event.Pointer.t -> unit) option

val onPointerDownGet : domProps -> (React.Event.Pointer.t -> unit) option

val onPointerMoveGet : domProps -> (React.Event.Pointer.t -> unit) option

val onPointerUpGet : domProps -> (React.Event.Pointer.t -> unit) option

val onPointerCancelGet : domProps -> (React.Event.Pointer.t -> unit) option

val onPointerOutGet : domProps -> (React.Event.Pointer.t -> unit) option

val onPointerLeaveGet : domProps -> (React.Event.Pointer.t -> unit) option

val onGotPointerCaptureGet : domProps -> (React.Event.Pointer.t -> unit) option

val onLostPointerCaptureGet : domProps -> (React.Event.Pointer.t -> unit) option

val onScrollGet : domProps -> (React.Event.UI.t -> unit) option

val onWheelGet : domProps -> (React.Event.Wheel.t -> unit) option

val onAbortGet : domProps -> (React.Event.Media.t -> unit) option

val onCanPlayGet : domProps -> (React.Event.Media.t -> unit) option

val onCanPlayThroughGet : domProps -> (React.Event.Media.t -> unit) option

val onDurationChangeGet : domProps -> (React.Event.Media.t -> unit) option

val onEmptiedGet : domProps -> (React.Event.Media.t -> unit) option

val onEncrypetdGet : domProps -> (React.Event.Media.t -> unit) option

val onEndedGet : domProps -> (React.Event.Media.t -> unit) option

val onErrorGet : domProps -> (React.Event.Media.t -> unit) option

val onLoadedDataGet : domProps -> (React.Event.Media.t -> unit) option

val onLoadedMetadataGet : domProps -> (React.Event.Media.t -> unit) option

val onLoadStartGet : domProps -> (React.Event.Media.t -> unit) option

val onPauseGet : domProps -> (React.Event.Media.t -> unit) option

val onPlayGet : domProps -> (React.Event.Media.t -> unit) option

val onPlayingGet : domProps -> (React.Event.Media.t -> unit) option

val onProgressGet : domProps -> (React.Event.Media.t -> unit) option

val onRateChangeGet : domProps -> (React.Event.Media.t -> unit) option

val onSeekedGet : domProps -> (React.Event.Media.t -> unit) option

val onSeekingGet : domProps -> (React.Event.Media.t -> unit) option

val onStalledGet : domProps -> (React.Event.Media.t -> unit) option

val onSuspendGet : domProps -> (React.Event.Media.t -> unit) option

val onTimeUpdateGet : domProps -> (React.Event.Media.t -> unit) option

val onVolumeChangeGet : domProps -> (React.Event.Media.t -> unit) option

val onWaitingGet : domProps -> (React.Event.Media.t -> unit) option

val onLoadGet : domProps -> (React.Event.Image.t -> unit) option

val onAnimationStartGet : domProps -> (React.Event.Animation.t -> unit) option

val onAnimationEndGet : domProps -> (React.Event.Animation.t -> unit) option

val onAnimationIterationGet : domProps -> (React.Event.Animation.t -> unit) option

val onTransitionEndGet : domProps -> (React.Event.Transition.t -> unit) option

val accentHeightGet : domProps -> string option

val accumulateGet : domProps -> string option

val additiveGet : domProps -> string option

val alignmentBaselineGet : domProps -> string option

val allowReorderGet : domProps -> string option

val alphabeticGet : domProps -> string option

val amplitudeGet : domProps -> string option

val arabicFormGet : domProps -> string option

val ascentGet : domProps -> string option

val attributeNameGet : domProps -> string option

val attributeTypeGet : domProps -> string option

val autoReverseGet : domProps -> string option

val azimuthGet : domProps -> string option

val baseFrequencyGet : domProps -> string option

val baseProfileGet : domProps -> string option

val baselineShiftGet : domProps -> string option

val bboxGet : domProps -> string option

val begin_Get : domProps -> string option

val biasGet : domProps -> string option

val byGet : domProps -> string option

val calcModeGet : domProps -> string option

val capHeightGet : domProps -> string option

val clipGet : domProps -> string option

val clipPathGet : domProps -> string option

val clipPathUnitsGet : domProps -> string option

val clipRuleGet : domProps -> string option

val colorInterpolationGet : domProps -> string option

val colorInterpolationFiltersGet : domProps -> string option

val colorProfileGet : domProps -> string option

val colorRenderingGet : domProps -> string option

val contentScriptTypeGet : domProps -> string option

val contentStyleTypeGet : domProps -> string option

val cursorGet : domProps -> string option

val cxGet : domProps -> string option

val cyGet : domProps -> string option

val dGet : domProps -> string option

val decelerateGet : domProps -> string option

val descentGet : domProps -> string option

val diffuseConstantGet : domProps -> string option

val directionGet : domProps -> string option

val displayGet : domProps -> string option

val divisorGet : domProps -> string option

val dominantBaselineGet : domProps -> string option

val durGet : domProps -> string option

val dxGet : domProps -> string option

val dyGet : domProps -> string option

val edgeModeGet : domProps -> string option

val elevationGet : domProps -> string option

val enableBackgroundGet : domProps -> string option

val end_Get : domProps -> string option

val exponentGet : domProps -> string option

val externalResourcesRequiredGet : domProps -> string option

val fillGet : domProps -> string option

val fillOpacityGet : domProps -> string option

val fillRuleGet : domProps -> string option

val filterGet : domProps -> string option

val filterResGet : domProps -> string option

val filterUnitsGet : domProps -> string option

val floodColorGet : domProps -> string option

val floodOpacityGet : domProps -> string option

val focusableGet : domProps -> string option

val fontFamilyGet : domProps -> string option

val fontSizeGet : domProps -> string option

val fontSizeAdjustGet : domProps -> string option

val fontStretchGet : domProps -> string option

val fontStyleGet : domProps -> string option

val fontVariantGet : domProps -> string option

val fontWeightGet : domProps -> string option

val fomatGet : domProps -> string option

val fromGet : domProps -> string option

val fxGet : domProps -> string option

val fyGet : domProps -> string option

val g1Get : domProps -> string option

val g2Get : domProps -> string option

val glyphNameGet : domProps -> string option

val glyphOrientationHorizontalGet : domProps -> string option

val glyphOrientationVerticalGet : domProps -> string option

val glyphRefGet : domProps -> string option

val gradientTransformGet : domProps -> string option

val gradientUnitsGet : domProps -> string option

val hangingGet : domProps -> string option

val horizAdvXGet : domProps -> string option

val horizOriginXGet : domProps -> string option

val ideographicGet : domProps -> string option

val imageRenderingGet : domProps -> string option

val in_Get : domProps -> string option

val in2Get : domProps -> string option

val interceptGet : domProps -> string option

val kGet : domProps -> string option

val k1Get : domProps -> string option

val k2Get : domProps -> string option

val k3Get : domProps -> string option

val k4Get : domProps -> string option

val kernelMatrixGet : domProps -> string option

val kernelUnitLengthGet : domProps -> string option

val kerningGet : domProps -> string option

val keyPointsGet : domProps -> string option

val keySplinesGet : domProps -> string option

val keyTimesGet : domProps -> string option

val lengthAdjustGet : domProps -> string option

val letterSpacingGet : domProps -> string option

val lightingColorGet : domProps -> string option

val limitingConeAngleGet : domProps -> string option

val localGet : domProps -> string option

val markerEndGet : domProps -> string option

val markerHeightGet : domProps -> string option

val markerMidGet : domProps -> string option

val markerStartGet : domProps -> string option

val markerUnitsGet : domProps -> string option

val markerWidthGet : domProps -> string option

val maskGet : domProps -> string option

val maskContentUnitsGet : domProps -> string option

val maskUnitsGet : domProps -> string option

val mathematicalGet : domProps -> string option

val modeGet : domProps -> string option

val numOctavesGet : domProps -> string option

val offsetGet : domProps -> string option

val opacityGet : domProps -> string option

val operatorGet : domProps -> string option

val orderGet : domProps -> string option

val orientGet : domProps -> string option

val orientationGet : domProps -> string option

val originGet : domProps -> string option

val overflowGet : domProps -> string option

val overflowXGet : domProps -> string option

val overflowYGet : domProps -> string option

val overlinePositionGet : domProps -> string option

val overlineThicknessGet : domProps -> string option

val paintOrderGet : domProps -> string option

val panose1Get : domProps -> string option

val pathLengthGet : domProps -> string option

val patternContentUnitsGet : domProps -> string option

val patternTransformGet : domProps -> string option

val patternUnitsGet : domProps -> string option

val pointerEventsGet : domProps -> string option

val pointsGet : domProps -> string option

val pointsAtXGet : domProps -> string option

val pointsAtYGet : domProps -> string option

val pointsAtZGet : domProps -> string option

val preserveAlphaGet : domProps -> string option

val preserveAspectRatioGet : domProps -> string option

val primitiveUnitsGet : domProps -> string option

val rGet : domProps -> string option

val radiusGet : domProps -> string option

val refXGet : domProps -> string option

val refYGet : domProps -> string option

val renderingIntentGet : domProps -> string option

val repeatCountGet : domProps -> string option

val repeatDurGet : domProps -> string option

val requiredExtensionsGet : domProps -> string option

val requiredFeaturesGet : domProps -> string option

val restartGet : domProps -> string option

val resultGet : domProps -> string option

val rotateGet : domProps -> string option

val rxGet : domProps -> string option

val ryGet : domProps -> string option

val scaleGet : domProps -> string option

val seedGet : domProps -> string option

val shapeRenderingGet : domProps -> string option

val slopeGet : domProps -> string option

val spacingGet : domProps -> string option

val specularConstantGet : domProps -> string option

val specularExponentGet : domProps -> string option

val speedGet : domProps -> string option

val spreadMethodGet : domProps -> string option

val startOffsetGet : domProps -> string option

val stdDeviationGet : domProps -> string option

val stemhGet : domProps -> string option

val stemvGet : domProps -> string option

val stitchTilesGet : domProps -> string option

val stopColorGet : domProps -> string option

val stopOpacityGet : domProps -> string option

val strikethroughPositionGet : domProps -> string option

val strikethroughThicknessGet : domProps -> string option

val stringGet : domProps -> string option

val strokeGet : domProps -> string option

val strokeDasharrayGet : domProps -> string option

val strokeDashoffsetGet : domProps -> string option

val strokeLinecapGet : domProps -> string option

val strokeLinejoinGet : domProps -> string option

val strokeMiterlimitGet : domProps -> string option

val strokeOpacityGet : domProps -> string option

val strokeWidthGet : domProps -> string option

val surfaceScaleGet : domProps -> string option

val systemLanguageGet : domProps -> string option

val tableValuesGet : domProps -> string option

val targetXGet : domProps -> string option

val targetYGet : domProps -> string option

val textAnchorGet : domProps -> string option

val textDecorationGet : domProps -> string option

val textLengthGet : domProps -> string option

val textRenderingGet : domProps -> string option

val to_Get : domProps -> string option

val transformGet : domProps -> string option

val u1Get : domProps -> string option

val u2Get : domProps -> string option

val underlinePositionGet : domProps -> string option

val underlineThicknessGet : domProps -> string option

val unicodeGet : domProps -> string option

val unicodeBidiGet : domProps -> string option

val unicodeRangeGet : domProps -> string option

val unitsPerEmGet : domProps -> string option

val vAlphabeticGet : domProps -> string option

val vHangingGet : domProps -> string option

val vIdeographicGet : domProps -> string option

val vMathematicalGet : domProps -> string option

val valuesGet : domProps -> string option

val vectorEffectGet : domProps -> string option

val versionGet : domProps -> string option

val vertAdvXGet : domProps -> string option

val vertAdvYGet : domProps -> string option

val vertOriginXGet : domProps -> string option

val vertOriginYGet : domProps -> string option

val viewBoxGet : domProps -> string option

val viewTargetGet : domProps -> string option

val visibilityGet : domProps -> string option

val widthsGet : domProps -> string option

val wordSpacingGet : domProps -> string option

val writingModeGet : domProps -> string option

val xGet : domProps -> string option

val x1Get : domProps -> string option

val x2Get : domProps -> string option

val xChannelSelectorGet : domProps -> string option

val xHeightGet : domProps -> string option

val xlinkActuateGet : domProps -> string option

val xlinkArcroleGet : domProps -> string option

val xlinkHrefGet : domProps -> string option

val xlinkRoleGet : domProps -> string option

val xlinkShowGet : domProps -> string option

val xlinkTitleGet : domProps -> string option

val xlinkTypeGet : domProps -> string option

val xmlnsGet : domProps -> string option

val xmlnsXlinkGet : domProps -> string option

val xmlBaseGet : domProps -> string option

val xmlLangGet : domProps -> string option

val xmlSpaceGet : domProps -> string option

val yGet : domProps -> string option

val y1Get : domProps -> string option

val y2Get : domProps -> string option

val yChannelSelectorGet : domProps -> string option

val zGet : domProps -> string option

val zoomAndPanGet : domProps -> string option

val aboutGet : domProps -> string option

val datatypeGet : domProps -> string option

val inlistGet : domProps -> string option

val prefixGet : domProps -> string option

val propertyGet : domProps -> string option

val resourceGet : domProps -> string option

val typeofGet : domProps -> string option

val vocabGet : domProps -> string option

val dangerouslySetInnerHTMLGet : domProps -> < __html : string > Js.t option

val suppressContentEditableWarningGet : domProps -> bool option

val suppressHydrationWarningGet : domProps -> bool option

val stringToComponent : string -> domProps React.component

val createElement : string -> ?props:domProps -> React.element array -> React.element

val createDOMElementVariadic : string -> ?props:domProps -> React.element array -> React.element

val jsxKeyed : string -> domProps -> ?key:string -> unit -> React.element

val jsx : string -> domProps -> React.element

val jsxs : string -> domProps -> React.element

val jsxsKeyed : string -> domProps -> ?key:string -> unit -> React.element